requirements.txt
streamlit==1.38.0
pandas==2.2.2
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
scikit-learn==1.5.1
